Added multiple stories :rocket:
