+++ __stories__/atoms/links.stories.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+import { Meta } from "@storybook/react";
+import { StandardInternalLink } from "../../src/components/design-system/atoms/standard-internal-link";
+import { StandardExternalLink } from "../../src/components/design-system/atoms/standard-external-link";
+
+export const Links: React.VFC = () => (
+  <>
+    <div>
+      <StandardInternalLink to={""}>An internal link</StandardInternalLink>
+    </div>
+    <div>
+      <StandardExternalLink>An external link</StandardExternalLink>
+    </div>
+  </>
+);
+
+export default {
+  title: "Atoms/Typography",
+  component: Links,
+} as Meta;
